+++ src/components/TextInput.js
@@ -1,7 +1,24 @@
-import React from 'react';
+import React, { useState } from 'react';
 
-const TextInput = () => {
-  return <div></div>;
+import TextField from '@material-ui/core/TextField';
+
+const TextInput = ({ label, handleOnChange }) => {
+  const [textValue, setTextValue] = useState('');
+
+  const onChange = e => {
+    setTextValue(e.target.value);
+    handleOnChange(e);
+  };
+
+  return (
+    <TextField
+      id="inputArray"
+      label={label}
+      variant="outlined"
+      value={textValue}
+      onChange={e => onChange(e)}
+    />
+  );
 };
 
 export default TextInput;
